back/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
pydantic==2.8.0
requests==2.32.0
